backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
openai==1.3.0
